STD: fix double promotion warnings

+++ lib/C/iec_std_FB.h
@@ -1602,7 +1602,7 @@ static void DERIVATIVE_body__(DERIVATIVE_data__ *data__) {
   // Initialise TEMP variables
 
   if (__GET_VAR(data__->RUN,)) {
-    __SET_VAR(data__->,XOUT,,((((3.0 * (__GET_VAR(data__->XIN,) - __GET_VAR(data__->X3,))) + __GET_VAR(data__->X1,)) - __GET_VAR(data__->X2,)) / (10.0 * ___TIME_TO_REAL((BOOL)__BOOL_LITERAL(TRUE),
+    __SET_VAR(data__->,XOUT,,(((((REAL)3.0 * (__GET_VAR(data__->XIN,) - __GET_VAR(data__->X3,))) + __GET_VAR(data__->X1,)) - __GET_VAR(data__->X2,)) / ((REAL)10.0 * ___TIME_TO_REAL((BOOL)__BOOL_LITERAL(TRUE),
       NULL,
       (TIME)__GET_VAR(data__->CYCLE,)))));
     __SET_VAR(data__->,X3,,__GET_VAR(data__->X2,));
